Add files via upload
Sum of Numbers - Python

Design a program with a loop using Python that asks the user to enter a series of positive numbers. The user should enter a negative number to signal the end of the series. After all the positive numbers have been entered, the program should display their sum.